chore(pos-app): pair token amount with ticker on transaction detail
The Asset received cell rendered the raw on-chain value (e.g. 10000)
when display.formatted was missing and never showed the ticker, which
violates the "always pair an amount with its asset" rule. Moves the
existing token formatter from printer.ts into currency.ts so both the
modal and the printer share one helper, and updates the modal to
prefer display.formatted, fall back to formatting value with
display.decimals, and always append display.assetSymbol.
