+++ dapps/pos-app/utils/currency.ts
@@ -105,3 +105,27 @@ export function formatFiatAmount(amount?: string, currency?: string): string {
 
   return formatAmountWithSymbol(formattedValue, currencyData);
 }
+
+const MAX_TOKEN_DECIMALS = 8;
+
+/**
+ * Formats a raw token amount (in smallest units) to a human-readable string.
+ * @param rawAmount - Token amount in smallest unit (e.g., "100000" for 0.0001 SOL).
+ * @param decimals - Token decimals (e.g., 9 for SOL, 6 for USDC).
+ * @returns Formatted amount string (e.g., "0.0001"). Caps at 8 decimal places.
+ */
+export function formatTokenAmount(rawAmount: string, decimals: number): string {
+  if (!rawAmount || decimals === 0) return rawAmount;
+
+  const padded = rawAmount.padStart(decimals + 1, "0");
+  const integerPart = padded.slice(0, -decimals) || "0";
+  const decimalPart = padded.slice(-decimals);
+
+  let trimmedDecimal = decimalPart.replace(/0+$/, "").padEnd(2, "0");
+
+  if (trimmedDecimal.length > MAX_TOKEN_DECIMALS) {
+    trimmedDecimal = trimmedDecimal.slice(0, MAX_TOKEN_DECIMALS);
+  }
+
+  return `${integerPart}.${trimmedDecimal}`;
+}
